Fix highline listing
